requirements.txt
streamlit
streamlit-multipage
streamlit-javascript
streamlit-authenticator
streamlit-wallet-connect
pdfkit
DateTime
pycoingecko
requests
yfinance
web3
lunarcrush
supabase
folium
geopy
plotly
streamlit-extras